module for json codecs
