+++ dynamodb/src/test/scala/zio/dynamodb/ItemJsonSerialisationSpec.scala
@@ -1,15 +1,14 @@
 package zio.dynamodb
 
+import zio.dynamodb.JsonCodec.Decoder._
 import zio.json._
 import zio.json.ast.Json
-import zio.Chunk
 import zio.test.ZIOSpecDefault
 import zio.Scope
 import zio.test.Spec
 import zio.test.TestEnvironment
 import zio.test.{ assert, assertTrue }
 import zio.test.Assertion.equalTo
-import scala.util.Try
 //import zio.prelude._
 
 object ItemJsonSerialisationSpec extends ZIOSpecDefault {
@@ -75,80 +74,7 @@ BS – Binary Set // TODO
   type Decoder[+A] = AttributeValue => Either[ItemError, A]
    */
 
-  def createMap(fields: Chunk[(String, Json)], map: AttributeValue.Map): Either[String, AttributeValue.Map] =
-    fields.toList match {
-      case Nil            =>
-        Right(map)
-      case (k, json) :: _ =>
-        decode(json) match {
-          case Right(av) => createMap(fields.tail, map + (k -> av))
-          case Left(err) => Left(err)
-        }
-    }
-
-  def decodeSS(xs: List[Json], acc: AttributeValue.StringSet): Either[String, AttributeValue.StringSet] =
-    xs match {
-      case Nil       => Right(acc)
-      case json :: _ =>
-        json match {
-          case Json.Str(s) =>
-            val x: AttributeValue.StringSet = acc + s
-            decodeSS(xs.tail, x)
-          case x           => Left(s"Invalid SS $x")
-        }
-    }
-
-  def decodeNS(xs: List[Json], acc: AttributeValue.NumberSet): Either[String, AttributeValue.NumberSet] =
-    xs match {
-      case Nil       => Right(acc)
-      case json :: _ =>
-        json match {
-          case Json.Str(s) =>
-            (acc + s).flatMap(decodeNS(xs.tail, _))
-          case x           => Left(s"Invalid SS $x")
-        }
-    }
-
-  def decodeL(xs: List[Json], acc: AttributeValue.List): Either[String, AttributeValue.List] =
-    xs match {
-      case Nil       => Right(acc)
-      case json :: _ =>
-        decode(json) match {
-          case Right(av) => decodeL(xs.tail, acc + av)
-          case Left(err) => Left(err)
-        }
-    }
-
-  def decode(json: Json): Either[String, AttributeValue] =
-    json match {
-      case Json.Obj(Chunk("N" -> Json.Str(d)))      =>
-        Try(BigDecimal(d)).fold(
-          _ => Left(s"Invalid Number $d"),
-          n => Right(AttributeValue.Number(n))
-        )
-      case Json.Obj(Chunk("S" -> Json.Str(s)))      => Right(AttributeValue.String(s))
-      // Note Json.Num is handles via Json.Str
-      case Json.Obj(Chunk("BOOL" -> Json.Bool(b)))  => Right(AttributeValue.Bool(b))
-      case Json.Obj(Chunk("L" -> Json.Arr(a)))      => decodeL(a.toList, AttributeValue.List.empty)
-      case Json.Obj(Chunk("SS" -> Json.Arr(chunk))) => decodeSS(chunk.toList, AttributeValue.StringSet.empty)
-      case Json.Obj(Chunk("NS" -> Json.Arr(chunk))) => decodeNS(chunk.toList, AttributeValue.NumberSet.empty)
-      case Json.Obj(Chunk("M" -> Json.Obj(fields))) => createMap(fields, AttributeValue.Map.empty)
-//      case Json.Obj(Chunk(_ -> a))              => Left(s"TODO ${a.getClass.getName} $a")
-
-      case Json.Obj(fields) if fields.isEmpty       => Left("empty AttributeValue Map found")
-      case Json.Obj(fields)                         => // returns an  AttributeValue.Map
-        createMap(fields, AttributeValue.Map.empty)
-      // for collections
-      case Json.Str(s)                              => Right(AttributeValue.String(s))
-      case Json.Bool(b)                             => Right(AttributeValue.Bool(b))
-      // Note Json.Num is handles via Json.Str
-      case a                                        => Left(s"Only top level objects are supported, found $a")
-    }
-
-  def toAttrMap(fields: List[(AttributeValue.String, AttributeValue)], acc: AttrMap): AttrMap =
-    fields.map { case (avStr -> av) => avStr.value -> av }.foldLeft(acc)(_ + _)
-
-  override def spec: Spec[TestEnvironment with Scope, Any]                                    =
+  override def spec: Spec[TestEnvironment with Scope, Any] =
     suite("ItemJsonSerialisationSpec")(
       test("decode top level map") {
         val s   =
@@ -313,7 +239,7 @@ BS – Binary Set // TODO
         val ast = "[]".fromJson[Json].getOrElse(Json.Null)
         assert(decode(ast))(
           equalTo(
-            Left("Only top level objects are supported, found []")
+            Left("top level arrays are not supported, found []")
           )
         )
       },
@@ -332,9 +258,9 @@ BS – Binary Set // TODO
         x match {
           case Right(AttributeValue.Map(map)) =>
             assertTrue(
-              toAttrMap(map.toList, AttrMap.empty) == AttrMap.empty + ("id" -> AttributeValue.String(
+              toAttrMap(map.toList) == AttrMap.empty + ("id" -> AttributeValue.String(
                 "101"
-              )) + ("count"                                                 -> AttributeValue.Number(BigDecimal(42)))
+              )) + ("count"                                  -> AttributeValue.Number(BigDecimal(42)))
             )
           case _                              => assertTrue(false)
         }
@@ -344,16 +270,16 @@ BS – Binary Set // TODO
         val s   =
           """{
               "foo": {
-                    "name": {
-                       "S": "Avi"
-                     }                    
+                  "name": {
+                      "S": "Avi"
+                    }                    
               }
           }"""
         val ast = s.fromJson[Json].getOrElse(Json.Null)
         val x   = decode(ast)
         x match {
           case Right(AttributeValue.Map(map)) =>
-            val translated = toAttrMap(map.toList, AttrMap.empty)
+            val translated = toAttrMap(map.toList)
             val nestedAv   = AttributeValue.Map.empty + ("name" -> AttributeValue.String("Avi"))
             val expected   = AttrMap.empty + ("foo"             -> nestedAv)
             assertTrue(translated == expected)
@@ -375,7 +301,7 @@ BS – Binary Set // TODO
         val x   = decode(ast)
         x match {
           case Right(AttributeValue.Map(map)) =>
-            val translated = toAttrMap(map.toList, AttrMap.empty)
+            val translated = toAttrMap(map.toList)
             val expected   = AttrMap.empty +
               ("id"     -> AttributeValue.String("101")) +
               ("nested" -> obj("bar")) +
